move links array creation test into test_colloid
