+++ tests/unit/tests.c
@@ -177,7 +177,6 @@ __host__ int tests_create(int argc, char ** argv) {
   test_fe_ternary_suite();
 
   /* Memory management */
-  test_create_links_arrays_suite();
   test_links_arrays_accessors_suite();
 
   return 0;
